Assets/requirements.txt
# Function dependencies

google-cloud-storage==1.31.2
google-cloud-asset==2.1.0
google-cloud-resource-manager==0.30.2
google-cloud-pubsub==1.7.0